Preserve floating-point precision when mapping areas to and from level 0
OpenSlide splits large regions into 4096-pixel wide or tall areas. When
moving to the next area, the advances in x and y are mapped to level 0
by multiplying with the downsample factor, and then the level 0
coordinates are mapped back to the region level by dividing with the
same downsample factor.

When the starting x, y are zeros, preserving all the precision prevents
integer coordinates become fractional after mapping, particularly when
vendors such as Aperio use non-integer downsample values.

When the starting x, y are non-zeros, preserving all the precision still
help cairo_translate() putting pixels at the correct position,
regardless whether the tx, ty being converted to integer at the final
translation.

+++ src/openslide-vendor-mirax.c
@@ -327,7 +327,7 @@ static bool read_tile(openslide_t *osr,
 }
 
 static bool paint_region(openslide_t *osr G_GNUC_UNUSED, cairo_t *cr,
-                         int64_t x, int64_t y,
+                         double x, double y,
                          struct _openslide_level *level,
                          int32_t w, int32_t h,
                          GError **err) {
